relation manager
a little bit complicated, filament provides a very beautiful UI for this relation manager. Note: don't forget to register RelationManager in the resource in getRelations method. The RelationManager is always created for the model you want to work with through connections. In this case, you want to manage posts related to the category. That is, you want to add the ability to edit, view and manage posts inside the category page. That's why you use the PostsRelationManager.