charts
So basically to add a new chart we need to use an artisan command: php artisan make:filament-widget. Then we can customize our widget in Filament/Widgets/NameOfYourChartWidget. For better data providing we can use a command composer require flowframe/laravel-trend to install a package that makes it easy to provide data to chart